Python Code for Daily-Dairy Billing
It's a code that first authenticates the user & then allows them to review rates and data of certain dairy stocks. Then it also empowers the user to connect with excel inside the python code instead of producing any data mayhem. With such a modest user-interactive window, the user could complete their job of taking orders from merchants of a dairy store with ease & comfort.
It was my first-ever success code in my life. #FirstCode